Cambios en la configuración del borrado según el usuario

+++ DeleteConcep.php
@@ -3,12 +3,20 @@
 require 'General.php';
 session_start();
 
+$user_session = General::getUserSession();
+
 $id = 0;
 if (isset($_GET['id'])) {
     $id = $_GET['id'];
 }
-
-if (General::deleteConcept($id)){
-    header("location:Adminis.php");
+if ($user_session['rol']=='ADMIN'){
+    if (General::updateConceptToDelete($id)){
+        header("location:Adminis.php");
+   }
+}else{
+   if (General::deleteConcept($id)){
+        header("location:Adminis.php");
+   }
 }
+
 ?>
